refactor(popover): menu directives
moves split button menu directives to popover and updates examples

closes #676
