+++ projects/cashmere/src/lib/button/index.ts
@@ -1,5 +1,4 @@
 export {AnchorComponent} from './anchor.component';
 export {ButtonComponent} from './button.component';
 export {ButtonModule} from './button.module';
-export {ButtonItemDirective} from './split-button/directives/button-item.directive';
 export {SplitButtonComponent, SplitButtonClickEvent} from './split-button/split-button.component';
