+++ projects/cashmere/src/lib/button/split-button/split-button.component.html
@@ -28,7 +28,7 @@
 </span>
 
 <hc-pop #splitMenu [xAlign]="menuPosition" [autoCloseOnContentClick]="autoCloseMenuOnClick" [showArrow]="false">
-    <div class="hc-menu-panel">
+    <div hcMenu>
         <ng-content select="[hcButtonItem]"></ng-content>
     </div>
 </hc-pop>
