+++ projects/cashmere/src/lib/pop/directives/menu.directive.ts
@@ -1,0 +1,10 @@
+import {Directive, HostBinding} from '@angular/core';
+
+/** The Menu directive provides a standard way of displaying a series of selectable elements in a popover */
+@Directive({
+    selector: '[hcMenu]'
+})
+export class MenuDirective {
+    @HostBinding('class.hc-menu-panel')
+    _hostClass = true;
+}
