+++ projects/cashmere/src/lib/pop/index.ts
@@ -1,3 +1,3 @@
-export {HcPopoverAnchorDirective} from './popover-anchor.directive';
+export {HcPopoverAnchorDirective} from './directives/popover-anchor.directive';
 export {PopModule} from './popover.module';
 export {HcPopComponent} from './popover.component';
